refactor[#523] : entities/marking/api 타입 리팩토링
- entities/marking/api 하위 훅들에서 많이 쓰이는 타입은 entities/marking/api/type.ts에 정의
