+++ src/entities/marking/api/getMyMakerList.ts
@@ -2,14 +2,9 @@ import { skipToken, useQuery } from "@tanstack/react-query";
 import { useTiling } from "@/entities/map/lib";
 import { apiClient } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
-import { MARKER_END_POINT, markerQueryKey } from "../constants";
-
-interface Marker {
-  markingId: number;
-  previewImage: string;
-  lat: number;
-  lng: number;
-}
+import { MARKER_END_POINT } from "../constants";
+import { markerQueryKey } from "./queryKey";
+import type { Marker } from "./type";
 
 type GetMyMarkerListResponse = Marker[];
